Import fenced effect manager banks
Imports ef/efmanager.c behind NDS_IMPORT_BATTLESHIP_EFFECT_MANAGER and keeps deferred particle visuals on diagnostic shims. Source path: efmanager.c:1734-1758 allocates the effect pool and loads EFCommonEffects1/2/3; efdisplay.c:26-34,84-105 installs display callbacks and the particle bank path; scvsbattlefiles.c:11-34 defines the VSBattle status-buffer envelope.

The effect-bank proof exposed reloc metadata saturation: the live file count reached 64, starving later natural throw motion/submotion loads. Raised NDS_RELOC_LOADED_FILE_CAPACITY to 96; mode 163 restored throwDmg=0->12 while reporting effect bank residency.

Ledger: before head207900 reloc653968 stage202816 fighter242480 if208672; fenced effect manager head240892 reloc748672 stage202816 fighter242480 if208672 other94704. The 128 KiB reserve remains intact.

Gates: verify-battle-playable-harness.ps1 -ImportBattleShipEffectManager -DelaySeconds 45; verify-dev-fast.ps1 -Build -DelaySeconds 3; verify-boundary.ps1 -DelaySeconds 3; check-harness-registry.ps1; check-docs.ps1.

+++ src/import/battleship_ifcommon.c
@@ -43,7 +43,7 @@ extern void gcAddAnimJointAll(GObj *gobj, AObjEvent32 **anim_joints,
                               f32 anim_frame);
 extern void gcPlayAnimAll(GObj *gobj);
 extern void gcSetupCustomDObjs(GObj *gobj, DObjDesc *dobjdesc, DObj **dobjs,
-                               u32 matrix_kind, u32 mobj_kind, u32 aobj_kind);
+                               u8 matrix_kind, u8 mobj_kind, u8 aobj_kind);
 extern void grWallpaperResumePerspUpdate(void);
 extern void grWallpaperPausePerspUpdate(void);
 extern void grWallpaperRunProcessAll(void);
